test: remove slow benchmarks
