+++ .github/workflows/test.yml
@@ -68,7 +68,7 @@ jobs:
         run: pip install -e '.[dev]'
       - name: Run tests
         run:
-          python -m pytest --cov-fail-under=98 -n=auto --cov=hotpdf --cov-report term-missing -m "not slow"
+          python -m pytest --cov-fail-under=98 -n=auto --cov=hotpdf --cov-report term-missing
       - name: Upload coverage to coveralls
         if: github.event_name == 'push'
         uses: coverallsapp/github-action@v2.2.3
