+++ pyproject.toml
@@ -89,7 +89,3 @@ strict = true
 [tool.pytest.ini_options]
 log_cli=true
 log_level="NOTSET"
-addopts="--cov=. --cov-report term-missing -n=auto"
-markers = [
-    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
-]
